notification some problem solv & notun kore korlam again but just group request send show done
